Se agrega la relacion a lotes

+++ services/sapadapter-service/Omicron.SapAdapter.Repository/Model/BusinessModels/AssignedBatches.cs
@@ -23,7 +23,7 @@ namespace Omicron.SapAdapter.Entities.Model.BusinessModels
         /// Gets or sets Code.
         /// </summary>
         /// <value>The code.</value>
-        public double CantidadSeleccionada { get; set; }
+        public decimal CantidadSeleccionada { get; set; }
 
         /// <summary>
         /// Gets or sets Code.
